Update rspec setup
This updates the rspec setup in the gem to follow our now standard pattern for rspec.

Any dependencies we use in rspec are defined in separate files stored in `spec/support`, including their config.

This then allows us to update `spec_helper.rb` to exactly match the spec helper in our other gems.

The aim is to simplify maintenance, adopt a consistent approach, and make it clear what additional dependencies we have in our tests.