Fix code coverage
PR #16 switched this project from CodeClimate to SonarCloud. When on CodeClimate our code coverage was 100% 🥳. After switching to SonarCloud its 0% 😱

After running it locally and checking what SimpleCov is reporting, it starts to make sense. SonarCloud we have learnt does not just take the final figure SimpleCov spits out. Instead it compares the details in `coverage/.resultset.json` with the code files to confirm what has been covered.

If you look at the detail of the SimpleCov report, its listing only the files in `spec/`! It shows there is no coverage for the actual code in `lib/`, but still reports 100%. Hence when SonarCloud analyses the results it comes back with a big fat 0.

This PR is about fixing the test coverage so we can report accurately what test coverage we have.