Update gem require to match other projects
The one difference I spotted between this project and our others is the way the files are required.

In this project it seems they are done in the gemspec. Everything works and the tests run so shouldn't be an issue. In all our other gems though we don't have this line in the gemspec, and its not what gets added if you use bundler to generate your initial gem project.

So as the first discrepancy we saw, this change updates the project to follow the convention of the others and instead require the `lib/` files in a support file which gets called when the tests run.

And voila! We now have accurate test coverage being reported 97.92% and our lib files are listed in the Simplecov report.