Switch upload to use aws:kms encryption
https://eaflood.atlassian.net/browse/RUBY-1099

The gem has alwasy encrypted on upload. This means irrespective of whether encryption at rest is enabled on the bucket, any file we upload is encrypted.

After a recent review and consultation with AWS, our web-ops team is intending to go through all AWS S3 buckets and set the policy that [all uploads are required to be encrypted using KMS](https://docs.aws.amazon.com/sdk-for-ruby/v3/developer-guide/s3-example-enforce-server-side-encryption.html).

We have tested with a bucket where this has been applied and the current upload still works fine. But to ensure there are no issues going forward this change switches the type of encryption we use on upload to KMS.

This change should not effect any services still using buckets without the policy. The key thing is whatever we upload is still encrypted. This will just mean when those services are updated they should not encounter any issues with their S3 uploads.